+++ GR_Settings.lua
@@ -597,7 +597,7 @@ ns.addonSettings = {
                             [14] = '14 days',
                             [30] = '30 days (1 month)',
                             [190] = '190 days (3 months)',
-                            [380] = '380 days (6 months)',
+                            [180] = '180 days (6 months)',
                         }
                     end,
                     set = function(_, val) ns.gSettings.antiSpamDays = tonumber(val) end,