+++ shared/GR_Dataset.lua
@@ -19,6 +19,9 @@ function ds:WhatsNew() -- What's new in the current version
                     Discord: https://discord.gg/ZtS6Q2sKRH
                 (or click on the icon in the top left corner)|r
 
+    |CFFFFFF00v3.0.31 Notes|r
+        - Changed the 6 months to 180 days.
+        (NOTE: If you were using the 6 months, you will need to update your settings.)
     |CFFFFFF00v3.0.30 Notes|r
         - Fixed issue with missing no guild link localization.
         - Updated alert to missing guild link on login.